webapp/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-docx>=1.0.0
python-pptx>=0.6.21
chardet>=5.0.0
openai>=1.0.0
